Realized that the boundary layer is essentially not doing anything at all because the measurement frequency is too low. Close to the switching surface, the nullifying feedback is really close to zero so the switching law dominates, and any voltage actuation that comes out of the control law is the 15 volt rails essentially. This is so large that by the next timestep, the switching law has evaluated to the opposite rail and the boundary layer is skipped over entirely. Essentially when I had my large boundary layer, I had a proportional controller on the S that was essentially utilizing some form of feedback linearization to produce a linear control law in S. Now that I have made it a smaller boundary layer, the sample frequency destroys the performance...unless I turn off the controller through my meta structure law - when the norm gets small enough, turn off the actuator. This actually ends up working almost perfectly and the pendulum lands right at the origin with some stiction switching only.
I think the next step would be to see what happens if I make the magnitude of my switching  law smaller. It is definitely going to change the acutaion because v_hat - v_s is going to be different. Currently that quantity keeps the acutation large at t=0 but cuts it down to about 3 or 4 volts pretty quickly. I don't know what is going to result here if I change it.